Multiple changes on deletion.

+++ corr-db/corrdb/common/managers/storage_manager.py
@@ -191,16 +191,17 @@ class StorageManager:
 
         for environment_id in project.history:
             _environment = EnvironmentModel.objects.with_id(environment_id)
-            if _environment.bundle["scope"] == "local":
-                result = self.storage_delete_file('bundle', _environment.bundle.location)
-                if result:
-                    logStat(deleted=True, bundle=_environment.bundle)
+            if _environment:
+                if _environment.bundle["scope"] == "local":
+                    result = self.storage_delete_file('bundle', _environment.bundle.location)
+                    if result:
+                        logStat(deleted=True, bundle=_environment.bundle)
+                        logStat(deleted=True, environment=_environment)
+                        _environment.bundle.delete()
+                        _environment.delete()
+                else:
                     logStat(deleted=True, environment=_environment)
-                    _environment.bundle.delete()
-                    _environment.delete()
-            else:
-                logStat(deleted=True, environment=_environment)
-            _environment.delete()
+                _environment.delete()
 
     def delete_record_files(self, record):
         """Delete a record files.
